pyproject.toml
[tool.poetry]
name = "simplify"
version = "0.1.0"
description = ""
authors = ["echapman2022 <ethan.chapman0@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8.5"
google-auth-oauthlib = "^0.4.2"
google-api-python-client = "^1.12.5"
google-auth-httplib2 = "^0.0.4"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
